requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.1.0
Django==4.1.6
django-celery-beat==2.4.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
flower==1.2.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.6.0
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
multidict==6.0.4
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.10.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==20.1
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
